+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog][] and this project adheres to
 ### Added
 
 - Support filtering feeds through an external command.
+- Opening informational text in digest email.
 
 ### Changed